lifted remaining cmd processing out of stocktopus and into slack pkg
