+++ pkg/slack/slack.go
@@ -4,13 +4,37 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"net/http"
 	"net/url"
+	"strconv"
 	"strings"
+	"time"
 
 	"github.com/go-redis/redis/v8"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/thorfour/stocktopus/pkg/stock"
 	"github.com/thorfour/stocktopus/pkg/stocktopus"
+)
+
+// Supported commands
+const (
+	addToList      = "WATCH"
+	printList      = "LIST"
+	removeFromList = "UNWATCH"
+	clear          = "CLEAR"
+	help           = "HELP"
+	infoCmd        = "INFO"
+	news           = "NEWS"
+	stats          = "STATS"
+
+	// Play money commands
+	buy       = "BUY"
+	sell      = "SELL"
+	deposit   = "DEPOSIT"
+	portfolio = "PORTFOLIO"
+	reset     = "RESET"
 )
 
 const (
@@ -26,7 +50,13 @@ type response struct {
 
 // SlashServer is a slack server that handles slash commands
 type SlashServer struct {
-	s *stocktopus.Stocktopus
+	s       *stocktopus.Stocktopus
+	cmdHist *prometheus.HistogramVec
+}
+
+// measureTime is a helper function to measure the execution time of a function
+func (s *SlashServer) measureTime(start time.Time, label string) {
+	s.cmdHist.WithLabelValues(label).Observe(time.Since(start).Seconds())
 }
 
 // New returns a new slash server
@@ -36,6 +66,12 @@ func New(kvstore *redis.Client, stocks stock.Lookup) *SlashServer {
 			KVStore:        kvstore,
 			StockInterface: stocks,
 		},
+		cmdHist: promauto.NewHistogramVec(prometheus.HistogramOpts{
+			Name: "command_timings",
+			Help: "A histogram of cmd request execution times",
+		},
+			[]string{"command"},
+		),
 	}
 }
 
@@ -65,7 +101,7 @@ func (s *SlashServer) Process(ctx context.Context, args url.Values) (string, err
 	}
 
 	text = strings.Split(strings.ToUpper(text[0]), " ")
-	return s.s.Command(ctx, text[0], text[1:], args)
+	return s.Command(ctx, text[0], text[1:], args)
 }
 
 // TODO determine ephermeralness of response
@@ -90,4 +126,121 @@ func (s *SlashServer) newResponse(resp http.ResponseWriter, message string, err 
 	resp.Header().Set("Content-Type", "application/json")
 	resp.Write(b)
 	return
+}
+
+// Command processes a stocktopus command
+func (s *SlashServer) Command(ctx context.Context, cmd string, args []string, info map[string][]string) (string, error) {
+	defer s.measureTime(time.Now(), cmd)
+
+	switch cmd {
+	case buy:
+		shares, err := strconv.Atoi(args[1])
+		if err != nil {
+			return "", err
+		}
+		a, err := s.s.Buy(ctx, args[0], uint64(shares), acctKey(info))
+		if err != nil {
+			return "", err
+		}
+
+		return a.String(), nil
+	case sell:
+		shares, err := strconv.Atoi(args[1])
+		if err != nil {
+			return "", err
+		}
+		a, err := s.s.Sell(ctx, args[0], uint64(shares), acctKey(info))
+		if err != nil {
+			return "", err
+		}
+
+		return a.String(), nil
+	case deposit:
+		amount, err := strconv.Atoi(args[0])
+		if err != nil {
+			return "", err
+		}
+		a, err := s.s.Deposit(ctx, float64(amount), acctKey(info))
+		if err != nil {
+			return "", err
+		}
+
+		return a.String(), nil
+	case portfolio:
+		a, err := s.s.Portfolio(ctx, acctKey(info))
+		if err != nil {
+			return "", err
+		}
+
+		a, err = s.s.Latest(ctx, a)
+		if err != nil {
+			return "", err
+		}
+
+		return a.String(), nil
+	case reset:
+		return "", s.s.Clear(ctx, acctKey(info))
+	case addToList:
+		return "", s.s.Add(ctx, args, listkey(args, info))
+	case printList:
+		acct, err := s.s.Print(ctx, listkey(args, info))
+		if err != nil {
+			return "", err
+		}
+
+		return acct.String(), nil
+	case removeFromList:
+		return "", s.s.Remove(ctx, args, listkey(args, info))
+	case clear:
+		return "", s.s.Clear(ctx, listkey(args, info))
+	case infoCmd:
+		c, err := s.s.Info(args[0])
+		if err != nil {
+			return "", err
+		}
+
+		return fmt.Sprintf("%v", c), nil // TODO
+	case news:
+		news, err := s.s.News(args[0])
+		if err != nil {
+			return "", err
+		}
+
+		return strings.Join(news, "\n"), nil
+	case stats:
+		stats, err := s.s.Stats(args[0])
+		if err != nil {
+			return "", err
+		}
+
+		return fmt.Sprintf("%v", stats), nil
+	default:
+		wl, err := s.s.GetQuotes(args)
+		if err != nil {
+			return "", err
+		}
+
+		return wl.String(), nil
+	}
+}
+
+func listkey(text []string, decodedMap url.Values) string {
+
+	// User and token to be used as watch list lookup
+	user := decodedMap["user_id"]
+	token := decodedMap["token"]
+
+	// If the first arg starts with '#' then it's the name of the list
+	if len(text) > 0 && strings.HasPrefix(text[0], "#") {
+		user = []string{strings.ToLower(text[0][1:]), decodedMap["team_id"][0]}
+	}
+
+	return fmt.Sprintf("%v%v", token, user)
+}
+
+func acctKey(decodedMap url.Values) string {
+	// User and token to be used as lookup
+	user := decodedMap["user_id"]
+	token := decodedMap["token"]
+	return fmt.Sprintf("%v%v%v", "ACCT", token, user)
 }
